requirements.txt
# =============================================================================
# Stock Trading Bot - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# TradingAgents Dependencies (from TradingAgents/requirements.txt)
# -----------------------------------------------------------------------------
typing-extensions
langchain-openai
langchain-experimental
langchain-anthropic
langchain-google-genai
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
rich
questionary

# -----------------------------------------------------------------------------
# Additional: Anthropic SDK
# -----------------------------------------------------------------------------
anthropic>=0.40.0

# -----------------------------------------------------------------------------
# Telegram Integration
# -----------------------------------------------------------------------------
python-telegram-bot>=21.0

# -----------------------------------------------------------------------------
# Configuration & Environment
# -----------------------------------------------------------------------------
pyyaml>=6.0
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# Logging
# -----------------------------------------------------------------------------
structlog>=24.0.0

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
